requirements.txt
asyncio
curses
windows-curses
python-dotenv